+++ uo_sys_framework/3DEN/display3DEN/Menu.hpp
@@ -9,5 +9,5 @@ class UO_FW_Module_Settings
 {
 	text = "Module Settings";
 	action = "edit3DENMissionAttributes 'UO_FW_Module_Settings';";
-	picture = "\a3\3DEN\Data\Cfg3DEN\History\makeNewLayer_ca.paa";
+	picture = "\a3\3DEN\Data\Displays\Display3DEN\PanelRight\modeModules_ca.paa";
 };