+++ uo_sys_framework/CoverMapModule/cfgFunctions.hpp
@@ -1,0 +1,10 @@
+class UO_FW_CoverMapFunctions {
+	file = "\x\UO_FW\addons\main\CoverMapModule\cfgFunctions\common";
+	class CoverMapLive {};
+	class CoverMapBriefing {};
+	class CoverMapInit {postInit = 1;};
+};
+class UO_FW_CoverMapModuleFunctions {
+	file = "\x\UO_FW\addons\main\CoverMapModule\cfgFunctions\modules";
+	class CoverMap {};
+};