+++ uo_sys_framework/GearModule/display3DEN/Menu.hpp
@@ -2,6 +2,5 @@ class UO_FW_Gear_Settings
 {
     text = "Configure Gear";
     action = "edit3DENMissionAttributes 'UO_FW_Gear_Settings';";
-    picture = "\a3\ui_f\data\IGUI\RscTitles\MPProgress\timer_ca.paa";
+    picture = "\a3\3DEN\Data\Displays\Display3DEN\EntityMenu\arsenal_ca.paa";
 };
-