+++ uo_sys_framework/MapCompassRemoverModule/cfgAttributeCategories/Object.hpp
@@ -1,8 +1,8 @@
-class UO_FW_MapCompassRemoverOptions 
+class UO_FW_MapCompassRemoverOptions
 {
-	displayName = "Remove Compass and Map Options"; 
+	displayName = "Remove Compass and Map Options";
 	collapsed = 1;
-	class Attributes 
+	class Attributes
 	{
 		class UO_FW_RemoveMap
 		{
@@ -10,7 +10,7 @@ class UO_FW_MapCompassRemoverOptions
 			tooltip = "If ticked the map for the unit will be removed after the briefing is done!";
 			property = "UO_FW_RemoveMap";
 			control = "Checkbox";
-			expression = "_this setVariable ['%s',_value,true];";
+			expression = "_this setVariable ['%s',_value,true]; if ((_value) && (UO_FW_SERVER_REMOVERMODULE_ALLOWED)) then { [{(time > 1)}, {_this unlinkItem 'ItemMap';}] call CBA_fnc_waitUntilAndExecute;};";
 			condition = "objectControllable";
 			defaultValue = "false";
 		};
@@ -19,6 +19,26 @@ class UO_FW_MapCompassRemoverOptions
 			displayName = "Remove compass after briefing";
 			tooltip = "If ticked the compass for the unit will be removed after the briefing is done!";
 			property = "UO_FW_RemoveCompass";
+			control = "Checkbox";
+			expression = "_this setVariable ['%s',_value,true]; if ((_value) && (UO_FW_SERVER_REMOVERMODULE_ALLOWED)) then { [{(time > 1)}, {_this unlinkItem 'ItemCompass';}] call CBA_fnc_waitUntilAndExecute;};";
+			condition = "objectControllable";
+			defaultValue = "false";
+		};
+		class UO_FW_KeepMapAI
+		{
+			displayName = "Keep map after briefing";
+			tooltip = "If ticked the map for the unit will be kept even if the module setting to remove all AI maps is enabled";
+			property = "UO_FW_KeepMapAI";
+			control = "Checkbox";
+			expression = "_this setVariable ['%s',_value,true];";
+			condition = "objectControllable";
+			defaultValue = "false";
+		};
+		class UO_FW_KeepCompassAI
+		{
+			displayName = "Keep compass after briefing";
+			tooltip = "If ticked the compass for the unit will be kept even if the module setting to remove all AI compasses is enabled";
+			property = "UO_FW_KeepCompassAI";
 			control = "Checkbox";
 			expression = "_this setVariable ['%s',_value,true];";
 			condition = "objectControllable";